스프링 부트와 AWS로 혼자 구현하는 웹 서비스/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/web/HelloController.java
package com.jojoldu.book.springboot.web;

import com.jojoldu.book.springboot.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController //json을 반환하는 컨트롤러로 생성
public class HelloController {

    @GetMapping("/hello")
    public String hello(){
        return "hello";
    }

    //RequestParam: 외부의 api로 넘긴 파라미터를 가져옴
    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name,
                                     @RequestParam("amount") int amount){
        return new HelloResponseDto(name, amount);
    }
}
